Cut up views.py
Views is getting too big and unmaintainable. It must be split up to increase efficiency.
