--- app/templates/app/new_map_good.html
@@ -1,82 +1,0 @@
-{% extends "app/base.html" %}
-{% load static %}
-{% load widget_tweaks %}
-{% block title %}New Map Test{% endblock %}
-{% block css %}
-<link rel="stylesheet" href="{% static "app/css/new_map.css " %}">
-<!-- <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script> -->
-{% endblock %}
-{% block content %}
-<br>
-<div id="container">
-    <!-- <select id="camSelect">
-        <option value="all">All Cameras</option>
-        <option value="good">Cameras in Good Condition</option>
-    </select>
-    <br> -->
-    <div id="map"></div>
-</div>
-<br>
-<div id="camInfo"></div>
-<script>
-function initMap(){
-    var center = {lat: 0, lng: 0};
-    var map = new google.maps.Map(document.getElementById('map'), {zoom: 2, center: center});
-    var infowindow = new google.maps.InfoWindow();
-    //var selector = document.getElementById('camSelect');
-    /*var markerArray = [];
-    selector.addEventListener("change", function() {
-        if (selector.value == "good"){
-            goodCAMs(map, infowindow, markerArray);
-        } else {
-            allCAMs(map, infowindow, markerArray);
-        }
-    }); */
-    //allCAMs(map, infowindow, markerArray);
-    //infowindow.setPosition({lat: -80, lng: 0});
-    {% for c in data %}
-    {% if c.is_active_video or c.is_active_image %}
-    var point = {lat: {{ c.latitude }}, lng: {{ c.longitude }}};
-    var marker = new google.maps.Marker({position: point, map: map, icon: 'https://storage.googleapis.com/support-kms-prod/SNP_2752125_en_v0'});
-    marker.addListener('click', function() {
-        infowindow.close();
-        var path = '<p>No Image</p>';
-        {% if c.type == 'ip' %}
-         {% if c.is_active_video %}
-         path = '<img src="http://{{ c.retrieval.ip }}{{ c.retrieval.video_path }}" alt="No Image" width="500">';
-         {% elif c.is_active_image %}
-         path = '<img src="http://{{ c.retrieval.ip }}{{ c.retrieval.image_path }}" alt="No Image" width="500">';
-         {% elif c.retrieval.video_path is not null %}
-         path = '<img src="http://{{ c.retrieval.ip }}{{ c.retrieval.video_path }}" alt="No Image" width="500">';
-         {% elif c.retrieval.image_path is not null %}
-         path = '<img src="http://{{ c.retrieval.ip }}{{ c.retrieval.image_path }}" alt="No Image" width="500">';
-         {% else %}
-         path = '<p>No Image</p>';
-         {% endif %}
-        {% elif c.type == 'non_ip' %}
-         {% if c.retrieval.snapshot_url is not null %}
-         path = '<img src="{{ c.retrieval.snapshot_url }}" alt="No Image" width="500">';
-         {% else %}
-         path = '<p>No Image</p>';
-         {% endif %}
-        {% elif c.type == 'stream' %}
-         {% if c.retrieval.m3u8_url is not null %}
-         path = '<img src="{{ c.retrieval.m3u8_url }}" alt="No Image" width="500">';
-         {% else %}
-         path = '<p>No Image</p>';
-         {% endif %}
-        {% endif %}
-        var place = '<div class="place" style="margin:auto;text-align:center;"><p>{% if c.city is not null %}{{ c.city }}, {% endif %}{% if c.state != "Null" %}{{ c.state }}, {% endif %}{{ c.country }}</p></div>';
-        var location = '<div class="loc" style="margin:auto;text-align:center;"><p>{{ c.latitude }}, {{ c.longitude }}</p></div>';
-        var contentString = path + place + location;
-        infowindow.setContent(contentString);
-        infowindow.setPosition({lat: {{ c.latitude }}, lng: {{ c.longitude }}});
-        infowindow.open(map);
-        //document.getElementById("camInfo").innerHTML=contentString;
-    });
-    {% endif %}
-    {% endfor %}
-}
-</script>
-<script asynch defer src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDICtCV5dWSi73-a7Bbhlw5Rs01Jax8EJY&callback=initMap"></script>
-{% endblock %}
