+++ app/urls.py
@@ -32,8 +32,6 @@ urlpatterns = [
     url(r'^password_reset/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
         auth_views.PasswordResetConfirmView.as_view(), name='password_reset_confirm'),
     url(r'^password_reset/complete/$', auth_views.PasswordResetCompleteView.as_view(), name='password_reset_complete'),
-    url(r'^new_map/$', app_views.new_map, name='new_map'),
-    url(r'^good_map/$', app_views.good_map, name='good_map'),
 
     #for testing API response
     url(r'^api_access/$', app_views.api_request, name='api_access'),
