+++ app/views/__init__.py
@@ -1,0 +1,5 @@
+from .static import *
+from .user import *
+from .cameras import cameras
+from .publications import publications
+from .team import team
